+++ packages/xchain-litecoin/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Update
 
 - Set Default network to `Network.Mainnet`
+- change contructor to accept NodeUrls = Record<Network, string>
 
 # v.0.10.3 (2022-10-04)